Modify the login interface
